Money/requirements.txt
flask==3.0.2
pandas==2.2.0
matplotlib==3.8.2
reportlab==4.1.0
openpyxl==3.1.2
flask-limiter==3.5.0
flask-wtf==1.2.1
psutil==5.9.5
flask-migrate==4.0.5
flask-caching==2.1.0
python-dotenv==1.0.1
gunicorn==21.2.0 